Dataset.tsx: move datasetAlerts logic to custom hook
